Change Types methods to use size_t

+++ include/exiv2/types.hpp
@@ -191,9 +191,9 @@ namespace Exiv2 {
      */
     struct EXIV2API DataBufRef {
         //! Constructor
-        explicit DataBufRef(std::pair<byte*, long> rhs) : p(rhs) {}
+        explicit DataBufRef(std::pair<byte*, size_t> rhs) : p(rhs) {}
         //! Pointer to a byte array and its size
-        std::pair<byte*, long> p;
+        std::pair<byte*, size_t> p;
     };
 
     /*!
@@ -209,9 +209,9 @@ namespace Exiv2 {
         //! Default constructor
         DataBuf();
         //! Constructor with an initial buffer size
-        explicit DataBuf(long size);
+        explicit DataBuf(size_t size);
         //! Constructor, copies an existing buffer
-        DataBuf(const byte* pData, long size);
+        DataBuf(const byte* pData, size_t size);
         /*!
           @brief Copy constructor. Transfers the buffer to the newly created
                  object similar to std::unique_ptr, i.e., the original object is
@@ -241,7 +241,7 @@ namespace Exiv2 {
                  buffer as a data pointer and size pair, resets the internal
                  buffer.
          */
-        EXV_WARN_UNUSED_RESULT std::pair<byte*, long> release();
+        EXV_WARN_UNUSED_RESULT std::pair<byte*, size_t> release();
 
          /*!
            @brief Free the internal buffer and reset the size to 0.
@@ -249,7 +249,7 @@ namespace Exiv2 {
         void free();
 
         //! Reset value
-        void reset(std::pair<byte*, long> =std::make_pair((byte*)(0),long(0)));
+        void reset(std::pair<byte*, size_t> =std::make_pair((byte*)(0),size_t(0)));
         //@}
 
         /*!
